Split legacy replay compatibility test
